fix: preserve 400 error classification in litellm compatibility mode
Native mode's maskedUpstreamErrorBody classifies a 400 into a short,
pre-vetted message plus (when available) the specific parameter that was
rejected — e.g. a rejected "logprobs" param surfaces as
param:"logprobs" instead of a generic "Invalid request". Gateways running
response_compatibility: litellm never saw any of this: normalizeError
unconditionally discarded the body (parameter was even named "_ []byte")
and replaced it with a hardcoded "Request failed" / param:null /
code:"<status>", regardless of what the router had already computed.

Extract the classification logic (previously private to proxy/errors.go)
into a new shared internal/upstreamerror package so both native mode and
litellm-compat mode run the identical classifier instead of one
discarding it. The classifier only ever returns a message from a small
fixed set plus a narrowly-restricted field/param name — never the
provider's own free-text message — so it's safe to run on a body that
hasn't been trusted yet (see TestTransformError /
TestClassifyBadRequest_NeverEchoesRawSensitiveText: raw provider bodies
with embedded credential names or internal URLs must never reach the
client verbatim, whether classified as 400 or not).

Re-classifying the router's own already-classified 400 output (which is
what litellm-compat mode's Transform actually receives — see
responseCompatibilityWriter.Close) needed one more fix to stay
idempotent: the "invalid_parameter" code (underscore) didn't match any
existing needle ("invalid parameter" has a space, "invalidparameter" has
none), so re-running the classifier on the router's own body fell through
to the generic default despite the original classification being correct.
Added "invalid_parameter" as an explicit needle.

internal/proxy/response_compat_test.go's
TestResponseCompatibilityWriterTransformsInternalError asserted the old
hardcoded generic shape for an unclassifiable body; updated to expect the
same "invalid_request" default native mode already produces for the same
input, instead of the old status-code-as-string "400".
